#38
- examples: мелкие правки
- tokens: добавлен JSDoc
- preset/bun: HandlerProvider перенесен в preset/server (patch)
