+++ examples/bun/src/components/Layout.tsx
@@ -1,6 +1,6 @@
-import type { ReactNode } from 'react';
+import type { CSSProperties, ReactNode } from 'react';
 
-const style = {
+const style: CSSProperties = {
   padding: '40px',
   margin: '0 auto',
   maxWidth: '960px',
