+++ src/preset/server/index.ts
@@ -4,3 +4,4 @@ export { getClientIp } from './utils/get-client-ip';
 export { getForwardedHeaders } from './utils/get-forwarded-headers';
 export { getHealthCheck } from './utils/get-health-check';
 export { getPageResponseFormat } from './utils/get-page-response-format';
+export { HandlerProvider } from './utils/handler-provider';
